rename connect.ts to components.ts
