+++ rollup.config.js
@@ -9,10 +9,10 @@ export default {
     index: 'src/index.ts',
     combineReducers: 'src/combineReducers.ts',
     compat: 'src/compat.ts',
+    components: 'src/components.ts',
     const: 'src/const.ts',
     createModel: 'src/createModel.ts',
     createStore: 'src/createStore.ts',
-    connect: 'src/connect.ts',
     devtools: 'src/devtools.ts',
     persist: 'src/persist.ts',
     routing: 'src/routingPlugin.ts',
