+++ typings/components.d.ts
